+++ server/boot/src/test/java/cn/torna/PasswordTest.java
@@ -1,7 +1,7 @@
 package cn.torna;
 
 import cn.torna.service.UserInfoService;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.util.DigestUtils;