+++ server/boot/src/test/java/cn/torna/TornaApplicationTests.java
@@ -1,11 +1,8 @@
 package cn.torna;
 
 import cn.torna.common.context.UserContext;
-import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
 
-@RunWith(SpringRunner.class)
 @SpringBootTest(classes = TornaApplication.class)
 class TornaApplicationTests {