+++ server/server-service/src/test/java/cn/torna/service/SwaggerParserTest.java
@@ -3,10 +3,9 @@ package cn.torna.service;
 import com.alibaba.fastjson.JSON;
 import io.swagger.parser.OpenAPIParser;
 import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.parser.core.models.ParseOptions;
 import io.swagger.v3.parser.core.models.SwaggerParseResult;
 import org.apache.commons.io.FileUtils;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.io.File;
 import java.io.IOException;