+++ server/server-web/src/main/java/cn/torna/web/interceptor/LoginInterceptor.java
@@ -10,7 +10,7 @@ import cn.torna.common.util.RequestUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.annotation.AnnotationUtils;
 import org.springframework.web.method.HandlerMethod;
-import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
+import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -19,10 +19,13 @@ import javax.servlet.http.HttpServletResponse;
  * @author tanghc
  */
 @Slf4j
-public class LoginInterceptor extends HandlerInterceptorAdapter {
+public class LoginInterceptor implements HandlerInterceptor {
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
+        if (!(handler instanceof HandlerMethod)) {
+            return true;
+        }
         HandlerMethod handlerMethod = (HandlerMethod) handler;
         NoLogin noLogin = handlerMethod.getMethodAnnotation(NoLogin.class);
         if (noLogin != null) {